Save uris
