+++ pkgs/record_use/lib/src/proto/usages_read.pbjson.dart
@@ -4,7 +4,7 @@
 //
 // @dart = 2.12
 
-// ignore_for_file: annotate_overrides, camel_case_types, comment_references
+// ignore_for_file: annotate_overrides, camel_case_types, comment_references, lines_longer_than_80_chars
 // ignore_for_file: constant_identifier_names, library_prefixes
 // ignore_for_file: non_constant_identifier_names, prefer_final_fields
 // ignore_for_file: unnecessary_import, unnecessary_this, unused_import
@@ -59,7 +59,7 @@ const Usage$json = {
       '3': 1,
       '4': 1,
       '5': 11,
-      '6': '.usages_shared.Definition',
+      '6': '.usages.Definition',
       '10': 'definition'
     },
     {
@@ -75,9 +75,8 @@ const Usage$json = {
 
 /// Descriptor for `Usage`. Decode as a `google.protobuf.DescriptorProto`.
 final $typed_data.Uint8List usageDescriptor = $convert.base64Decode(
-    'CgVVc2FnZRI5CgpkZWZpbml0aW9uGAEgASgLMhkudXNhZ2VzX3NoYXJlZC5EZWZpbml0aW9uUg'
-    'pkZWZpbml0aW9uEjEKCnJlZmVyZW5jZXMYAiADKAsyES51c2FnZXMuUmVmZXJlbmNlUgpyZWZl'
-    'cmVuY2Vz');
+    'CgVVc2FnZRIyCgpkZWZpbml0aW9uGAEgASgLMhIudXNhZ2VzLkRlZmluaXRpb25SCmRlZmluaX'
+    'Rpb24SMQoKcmVmZXJlbmNlcxgCIAMoCzIRLnVzYWdlcy5SZWZlcmVuY2VSCnJlZmVyZW5jZXM=');
 
 @$core.Deprecated('Use locationDescriptor instead')
 const Location$json = {
@@ -137,3 +136,56 @@ final $typed_data.Uint8List referenceDescriptor = $convert.base64Decode(
     '9uEiEKDGxvYWRpbmdfdW5pdBgCIAEoCVILbG9hZGluZ1VuaXQSOAoJYXJndW1lbnRzGAMgASgL'
     'MhgudXNhZ2VzX3NoYXJlZC5Bcmd1bWVudHNIAFIJYXJndW1lbnRzEi8KBmZpZWxkcxgEIAEoCz'
     'IVLnVzYWdlc19zaGFyZWQuRmllbGRzSABSBmZpZWxkc0ILCglyZWZlcmVuY2U=');
+
+@$core.Deprecated('Use definitionDescriptor instead')
+const Definition$json = {
+  '1': 'Definition',
+  '2': [
+    {
+      '1': 'identifier',
+      '3': 1,
+      '4': 1,
+      '5': 11,
+      '6': '.usages.Identifier',
+      '10': 'identifier'
+    },
+    {'1': 'line', '3': 2, '4': 1, '5': 13, '10': 'line'},
+    {'1': 'column', '3': 3, '4': 1, '5': 13, '10': 'column'},
+    {
+      '1': 'loading_unit',
+      '3': 4,
+      '4': 1,
+      '5': 9,
+      '9': 0,
+      '10': 'loadingUnit',
+      '17': true
+    },
+  ],
+  '8': [
+    {'1': '_loading_unit'},
+  ],
+};
+
+/// Descriptor for `Definition`. Decode as a `google.protobuf.DescriptorProto`.
+final $typed_data.Uint8List definitionDescriptor = $convert.base64Decode(
+    'CgpEZWZpbml0aW9uEjIKCmlkZW50aWZpZXIYASABKAsyEi51c2FnZXMuSWRlbnRpZmllclIKaW'
+    'RlbnRpZmllchISCgRsaW5lGAIgASgNUgRsaW5lEhYKBmNvbHVtbhgDIAEoDVIGY29sdW1uEiYK'
+    'DGxvYWRpbmdfdW5pdBgEIAEoCUgAUgtsb2FkaW5nVW5pdIgBAUIPCg1fbG9hZGluZ191bml0');
+
+@$core.Deprecated('Use identifierDescriptor instead')
+const Identifier$json = {
+  '1': 'Identifier',
+  '2': [
+    {'1': 'uri', '3': 1, '4': 1, '5': 9, '10': 'uri'},
+    {'1': 'parent', '3': 2, '4': 1, '5': 9, '9': 0, '10': 'parent', '17': true},
+    {'1': 'name', '3': 3, '4': 1, '5': 9, '10': 'name'},
+  ],
+  '8': [
+    {'1': '_parent'},
+  ],
+};
+
+/// Descriptor for `Identifier`. Decode as a `google.protobuf.DescriptorProto`.
+final $typed_data.Uint8List identifierDescriptor = $convert.base64Decode(
+    'CgpJZGVudGlmaWVyEhAKA3VyaRgBIAEoCVIDdXJpEhsKBnBhcmVudBgCIAEoCUgAUgZwYXJlbn'
+    'SIAQESEgoEbmFtZRgDIAEoCVIEbmFtZUIJCgdfcGFyZW50');
