+++ pkgs/record_use/lib/src/proto/usages_shared.pbjson.dart
@@ -38,60 +38,6 @@ final $typed_data.Uint8List metadataDescriptor = $convert.base64Decode(
     'CghNZXRhZGF0YRIdCgdjb21tZW50GAEgASgJSABSB2NvbW1lbnSIAQESGAoHdmVyc2lvbhgCIA'
     'EoCVIHdmVyc2lvbkIKCghfY29tbWVudA==');
 
-@$core.Deprecated('Use definitionDescriptor instead')
-const Definition$json = {
-  '1': 'Definition',
-  '2': [
-    {
-      '1': 'identifier',
-      '3': 1,
-      '4': 1,
-      '5': 11,
-      '6': '.usages_shared.Identifier',
-      '10': 'identifier'
-    },
-    {'1': 'line', '3': 2, '4': 1, '5': 13, '10': 'line'},
-    {'1': 'column', '3': 3, '4': 1, '5': 13, '10': 'column'},
-    {
-      '1': 'loading_unit',
-      '3': 4,
-      '4': 1,
-      '5': 9,
-      '9': 0,
-      '10': 'loadingUnit',
-      '17': true
-    },
-  ],
-  '8': [
-    {'1': '_loading_unit'},
-  ],
-};
-
-/// Descriptor for `Definition`. Decode as a `google.protobuf.DescriptorProto`.
-final $typed_data.Uint8List definitionDescriptor = $convert.base64Decode(
-    'CgpEZWZpbml0aW9uEjkKCmlkZW50aWZpZXIYASABKAsyGS51c2FnZXNfc2hhcmVkLklkZW50aW'
-    'ZpZXJSCmlkZW50aWZpZXISEgoEbGluZRgCIAEoDVIEbGluZRIWCgZjb2x1bW4YAyABKA1SBmNv'
-    'bHVtbhImCgxsb2FkaW5nX3VuaXQYBCABKAlIAFILbG9hZGluZ1VuaXSIAQFCDwoNX2xvYWRpbm'
-    'dfdW5pdA==');
-
-@$core.Deprecated('Use identifierDescriptor instead')
-const Identifier$json = {
-  '1': 'Identifier',
-  '2': [
-    {'1': 'uri', '3': 1, '4': 1, '5': 9, '10': 'uri'},
-    {'1': 'parent', '3': 2, '4': 1, '5': 9, '9': 0, '10': 'parent', '17': true},
-    {'1': 'name', '3': 3, '4': 1, '5': 9, '10': 'name'},
-  ],
-  '8': [
-    {'1': '_parent'},
-  ],
-};
-
-/// Descriptor for `Identifier`. Decode as a `google.protobuf.DescriptorProto`.
-final $typed_data.Uint8List identifierDescriptor = $convert.base64Decode(
-    'CgpJZGVudGlmaWVyEhAKA3VyaRgBIAEoCVIDdXJpEhsKBnBhcmVudBgCIAEoCUgAUgZwYXJlbn'
-    'SIAQESEgoEbmFtZRgDIAEoCVIEbmFtZUIJCgdfcGFyZW50');
-
 @$core.Deprecated('Use fieldsDescriptor instead')
 const Fields$json = {
   '1': 'Fields',
