+++ pkgs/record_use/lib/src/proto/usages_shared.proto
@@ -6,19 +6,6 @@ message Metadata {
   string version = 2;
 }
 
-message Definition {
-  Identifier identifier = 1;
-  uint32 line = 2;
-  uint32 column = 3;
-  optional string loading_unit = 4; // Optional, as some loading units are null
-}
-
-message Identifier {
-  string uri = 1; 
-  optional string parent = 2;
-  string name = 3;
-}
-
 message Fields {
   repeated Field fields = 1;
 }
