+++ pkgs/record_use/lib/src/proto/usages_storage.pbjson.dart
@@ -31,7 +31,7 @@ const Usages$json = {
       '3': 3,
       '4': 3,
       '5': 11,
-      '6': '.usages_shared.Definition',
+      '6': '.usages.Definition',
       '10': 'definitions'
     },
     {
@@ -56,10 +56,9 @@ const Usages$json = {
 /// Descriptor for `Usages`. Decode as a `google.protobuf.DescriptorProto`.
 final $typed_data.Uint8List usagesDescriptor = $convert.base64Decode(
     'CgZVc2FnZXMSMwoIbWV0YWRhdGEYASABKAsyFy51c2FnZXNfc2hhcmVkLk1ldGFkYXRhUghtZX'
-    'RhZGF0YRISCgR1cmlzGAIgAygJUgR1cmlzEjsKC2RlZmluaXRpb25zGAMgAygLMhkudXNhZ2Vz'
-    'X3NoYXJlZC5EZWZpbml0aW9uUgtkZWZpbml0aW9ucxIjCgVjYWxscxgEIAMoCzINLnVzYWdlcy'
-    '5Vc2FnZVIFY2FsbHMSKwoJaW5zdGFuY2VzGAUgAygLMg0udXNhZ2VzLlVzYWdlUglpbnN0YW5j'
-    'ZXM=');
+    'RhZGF0YRISCgR1cmlzGAIgAygJUgR1cmlzEjQKC2RlZmluaXRpb25zGAMgAygLMhIudXNhZ2Vz'
+    'LkRlZmluaXRpb25SC2RlZmluaXRpb25zEiMKBWNhbGxzGAQgAygLMg0udXNhZ2VzLlVzYWdlUg'
+    'VjYWxscxIrCglpbnN0YW5jZXMYBSADKAsyDS51c2FnZXMuVXNhZ2VSCWluc3RhbmNlcw==');
 
 @$core.Deprecated('Use usageDescriptor instead')
 const Usage$json = {
@@ -140,3 +139,56 @@ final $typed_data.Uint8List referenceDescriptor = $convert.base64Decode(
     '9uEiEKDGxvYWRpbmdfdW5pdBgCIAEoCVILbG9hZGluZ1VuaXQSOAoJYXJndW1lbnRzGAMgASgL'
     'MhgudXNhZ2VzX3NoYXJlZC5Bcmd1bWVudHNIAFIJYXJndW1lbnRzEi8KBmZpZWxkcxgEIAEoCz'
     'IVLnVzYWdlc19zaGFyZWQuRmllbGRzSABSBmZpZWxkc0ILCglyZWZlcmVuY2U=');
+
+@$core.Deprecated('Use definitionDescriptor instead')
+const Definition$json = {
+  '1': 'Definition',
+  '2': [
+    {
+      '1': 'identifier',
+      '3': 1,
+      '4': 1,
+      '5': 11,
+      '6': '.usages.Identifier',
+      '10': 'identifier'
+    },
+    {'1': 'line', '3': 2, '4': 1, '5': 13, '10': 'line'},
+    {'1': 'column', '3': 3, '4': 1, '5': 13, '10': 'column'},
+    {
+      '1': 'loading_unit',
+      '3': 4,
+      '4': 1,
+      '5': 9,
+      '9': 0,
+      '10': 'loadingUnit',
+      '17': true
+    },
+  ],
+  '8': [
+    {'1': '_loading_unit'},
+  ],
+};
+
+/// Descriptor for `Definition`. Decode as a `google.protobuf.DescriptorProto`.
+final $typed_data.Uint8List definitionDescriptor = $convert.base64Decode(
+    'CgpEZWZpbml0aW9uEjIKCmlkZW50aWZpZXIYASABKAsyEi51c2FnZXMuSWRlbnRpZmllclIKaW'
+    'RlbnRpZmllchISCgRsaW5lGAIgASgNUgRsaW5lEhYKBmNvbHVtbhgDIAEoDVIGY29sdW1uEiYK'
+    'DGxvYWRpbmdfdW5pdBgEIAEoCUgAUgtsb2FkaW5nVW5pdIgBAUIPCg1fbG9hZGluZ191bml0');
+
+@$core.Deprecated('Use identifierDescriptor instead')
+const Identifier$json = {
+  '1': 'Identifier',
+  '2': [
+    {'1': 'uri', '3': 1, '4': 1, '5': 13, '10': 'uri'},
+    {'1': 'parent', '3': 2, '4': 1, '5': 9, '9': 0, '10': 'parent', '17': true},
+    {'1': 'name', '3': 3, '4': 1, '5': 9, '10': 'name'},
+  ],
+  '8': [
+    {'1': '_parent'},
+  ],
+};
+
+/// Descriptor for `Identifier`. Decode as a `google.protobuf.DescriptorProto`.
+final $typed_data.Uint8List identifierDescriptor = $convert.base64Decode(
+    'CgpJZGVudGlmaWVyEhAKA3VyaRgBIAEoDVIDdXJpEhsKBnBhcmVudBgCIAEoCUgAUgZwYXJlbn'
+    'SIAQESEgoEbmFtZRgDIAEoCVIEbmFtZUIJCgdfcGFyZW50');
