+++ pkgs/record_use/lib/src/proto/usages_storage.proto
@@ -6,7 +6,7 @@ import "usages_shared.proto";
 message Usages {
   usages_shared.Metadata metadata = 1;
   repeated string uris = 2;
-  repeated usages_shared.Definition definitions = 3;
+  repeated Definition definitions = 3;
   repeated Usage calls = 4;
   repeated Usage instances = 5;
 }
@@ -29,4 +29,17 @@ message Reference {
     usages_shared.Arguments arguments = 3;
     usages_shared.Fields fields = 4;
   }
+}
+
+message Definition {
+  Identifier identifier = 1;
+  uint32 line = 2;
+  uint32 column = 3;
+  optional string loading_unit = 4; // Optional, as some loading units are null
+}
+
+message Identifier {
+  uint32 uri = 1; 
+  optional string parent = 2;
+  string name = 3;
 }
